requirements.txt
streamlit
qrcode
opencv-python
pycryptodome
